Add framework-specific code generation for AppKit, UIKit, and SwiftUI.
This kind of reverts the first commit in this series, which enabled multiple --framework options.
This would lead to name clashes. I think it's better to just call Shark multiple
times with varying parameters, if you really want that.
As an added bonus, we consider `.uikit` as the default and don't have to change the
CLI.