Roll back the default localization builder w/ LocalizedStringKey and provide an extension on String instead.
Although LocalizedStringKey seems to be a very important data type for SwiftUI entities,
there are still a lot of places where a plain String instead is used.
To cope for the need of LocalizedStringKey (which, e.g., is necessary to make automatic
MarkDown conversion work), we now provide an extension on String instead.